// src/types/jobData.ts
export type JobData = {
    title: string;
    company: string;
    description: string;
    tags: string[];
    tagLabel: string;
    tagVariant: "danger" | "success" | "info" | "default";
    postedAgo: string;
    salary: string;
    workType: string;
    companyDescription: string;
    jobDescription: string;
    responsibilities: string[];
    qualifications: string[];
};